Finalize text.ts refactoring
