+++ index.html
@@ -9,18 +9,17 @@
         Scripts
          -->
             <!-- Utilities -->
-        <script src="resources/build/utils.js" type="text/javascript"></script>
-        <script src="resources/build/moment.min.js" type="text/javascript"></script>
+        <script src="resources/build/utils/utils.js" type="text/javascript"></script>
+        <script src="resources/build/utils/loadData.js" type="text/javascript"></script>
+        <script src="resources/build/utils/relationshipCalc.js" type="text/javascript"></script>
             <!-- Graphics and drawing -->
         <script src="resources/build/graphics/text.js" type="text/javascript"></script>
-        <script src="resources/build/graphics/anim.js" type="text/javascript"></script>
-        <script src="resources/build/graphics/draw.js" type="text/javascript"></script>
+        <script src="resources/build/graphics/drawLines.js" type="text/javascript"></script>
             <!-- Structural -->
         <script src="resources/build/structure/nodes.js" type="text/javascript"></script>
         <script src="resources/build/structure/trees.js" type="text/javascript"></script>
         <script src="resources/build/structure/view.js" type="text/javascript"></script>
-            <!-- Load data -->
-        <script src="resources/build/loadData.js" type="text/javascript"></script>
+        
             <!-- Main -->
         <script src="resources/build/main.js" type="text/javascript"></script>
