+++ resources/ts/graphics/text.ts
@@ -1,139 +1,169 @@
-// Text attributes
-class TextAttr {
-  size: number;
-  prevHeight: number;
-  font: string;
-  style: string;
-  color: string;
-  baseSize: number;
+/**
+ * Represents a group of identically-styled segments of text.
+ */
+class StyledText {
+  // The texts
+  text: string[];
+  // Their attributes
+  attributes: TextAttributes;
 
-  constructor(_size: number, _font: string, _style: string, _color: string, _baseSize: number) {
-    this.size = _size;
+  /**
+   * Creates a new StyledText instance.
+   * @param text      The array of text
+   * @param attributes  The text attributes
+   */
+  constructor(text: string[], attributes: TextAttributes) {
+    this.text = text;
+    this.attributes = attributes;
+  }
+}
+
+
+/**
+ * Attributes of a segment of text.
+ */
+class TextAttributes {
+  // Sizing
+  size: number;
+  baseSize: number;
+  // Style
+  font: string;
+  color: string;
+  // Cached height
+  prevHeight: number;
+
+
+  /**
+   * Creates a TextAttributes instance.
+   * @param font      The font of the text
+   * @param color     The color of the text
+   * @param baseSize  The base size of the text
+   */
+  constructor(font: string, color: string, baseSize: number) {
+    this.size = baseSize;
+    this.baseSize = baseSize;
+    this.font = font;
+    this.color = color;
     this.prevHeight = 0;
-    this.font = _font;
-    this.style = _style;
-    this.color = _color;
-    this.baseSize = _baseSize;
   }
 
-  // Calculates the true font height on the canvas
-  calcFontHeight(fontStyle: string) {
+  /**
+   * Calculates the true size of text in this font on the canvas.
+   */
+  calcTextHeight(): number {
     let body = document.getElementsByTagName("body")[0];
     let tempDiv = document.createElement("div");
     let tempText = document.createTextNode("DUMMY TEXT");
 
     tempDiv.appendChild(tempText);
-    tempDiv.setAttribute("style", fontStyle);
+    // Set the text-style attributes of this div
+    tempDiv.setAttribute("style", `font: ${this.attributeString()}`);
 
     body.appendChild(tempDiv);
     let height = tempDiv.offsetHeight;
     body.removeChild(tempDiv);
 
-    // Remove the references
-    // @ts-ignore
-    body = null;
-    // @ts-ignore
+    // Now that we've done the work, we null tempDiv to remove the reference
+    // @ts-ignore (tempDiv isn't of type null, but I'm discarding of it)
     tempDiv = null;
-    // @ts-ignore
-    tempText = null;
 
     return height;
   }
 
-  getBaseSize() {
-    return this.baseSize;
-  }
-
-  getSize() {
-    return this.size;
-  }
-
-  setSize(newSize: number) {
-    this.size = newSize;
+  /**
+   * Scales the size of this TextAttribute (new size = base size * scale).
+   * @param scale The scale factor by which we scale the size.
+   */
+  scaleSize(scale: number) {
+    this.size = this.baseSize * scale;
     this.prevHeight = 0; // We have to recalculate the height
   }
+
+  /**
+   * Gets the line-height of text in with these attributes
+   */
   getHeight() {
     if (this.prevHeight <= 0) { // In this case, we haven't calculated it yet
-      this.prevHeight = this.calcFontHeight("font: " + this.asText());
+      this.prevHeight = this.calcTextHeight();
     }
     return this.prevHeight;
   }
 
-  asText() {
-    return this.style + " " + this.size + "px " + this.font;
+  /**
+   * Returns the attributes of this TextAttr in a CSS-styled string.
+   */
+  attributeString() {
+    return `${this.size}px ${this.font}`;
   }
 
-  // TODO - what element shoud this be?
-  apply(elem: any) {
-    // set the style
-    elem.context.font = this.asText();
-    // set the color
-    elem.context.fillStyle = this.color;
+  /**
+   * Applies the text-style to the given View.
+   * @param view  The View to which we apply these text attributes
+   */
+  apply(view: View) {
+    if (view.context !== null) {
+      view.context.font = this.attributeString();
+      view.context.fillStyle = this.color;
+    }
   }
 }
 
 
 //// FONTS
-const baseFontSize = 13;
-const detailFontSize = 11;
-const baseFont = new TextAttr(baseFontSize * scale, "sans-serif", "normal",
-    "#000000", baseFontSize);
-const detailFont = new TextAttr(detailFontSize * scale, "sans-serif", "normal",
-    "#333", detailFontSize);
+const baseFont = new TextAttributes("sans-serif", "#000", 13);
+const detailFont = new TextAttributes("sans-serif", "#333", 11);
 
 
-// TODO - is this ever anythig but a string?
-function parseTextAsArray(texts: any[]) {
-  let newText = [];
-  for (let text of texts) {
-    if (typeof text === "string") {
-      let s = text.split("\n");
-
-      for (let j = 0; j < s.length; j++) {
-        newText.push(s[j]);
-        if (j < s.length - 1) {
-          newText.push("\n");
-        }
-      }
-    } else {
-      newText.push(text);
-    }
+/**
+ * Renders all of the texts in a given array of StyledTexts.
+ * @param texts The StyledTexts which we are rendering
+ * @param view  The View on which the text will be rendered
+ * @param _x    The initial top-left x coordinate of the area on which we render the text
+ * @param _y    The initial top-left y coordinate
+ * @param display Whether this text should be displayed or not
+ *                    The default is true; if false, the text is likely only begin used for
+ *                    measurements.
+ */
+function renderText(texts: StyledText[], view: View, _x: number, _y: number,
+                    display: boolean = true): number[] | void {
+  if (view.context === null) {
+    return;
   }
 
-  return newText;
-}
-// TOdo canvasview
-function renderText(text: any[], view: any, _x: number, _y: number, real: boolean) {
-  let lastFont = baseFont;
   view.context.textBaseline = "top";
   let maxWidth = 0;
-  let maxLineHeight = 0;
+  let lineHeight = 0;
 
-  let x = _x;
-  let y = _y;
+  let cur_y = _y;
 
-  parseTextAsArray(text).map(e => {
-    if (typeof e == "string") {
-      if (e == "\n") {
-        x = _x // Reset X
-        y += maxLineHeight; // Add to the max height
-        maxLineHeight = 0; // reset line height
+
+  texts.map(t => {
+    // Apply the attributes of this StyledText
+    t.attributes.apply(view);
+
+    // Render all of the texts
+    for (let text of t.text) {
+      if (view.context === null) {
+        return;
       }
-      else { // just show the text
-        // If this is "real" text, we want to display it. If not real, that means we probably
-        // did this to calculate dimensions, so we don't want to show it.
-        if (real) {
-          view.context.fillText(e, x, y);
-        }
 
-        x += view.context.measureText(e).width;
-        maxWidth = Math.max(maxWidth, x - _x);
-        maxLineHeight = Math.max(maxLineHeight, lastFont.getHeight());
+      // Only draw it if it should be displayed
+      if (display)  {
+        view.context.fillText(text, _x, cur_y);
       }
-    } else { // it is a TextAttr
-      lastFont = e;
-      e.apply(view);
+
+      let curWidth = view.context.measureText(text).width - _x;
+      maxWidth = Math.max(maxWidth, curWidth);
+      lineHeight = Math.max(lineHeight, t.attributes.getHeight());
+
+
+      // Create a line-break between each one
+      cur_y += lineHeight;
+      lineHeight = 0;  // Reset the line-height
     }
+
   })
-  return [maxWidth, (y + maxLineHeight) - _y];
+
+  // Returns the width and height of the resulting text
+  return [maxWidth, (cur_y + lineHeight) - _y];
 }
